feat(product): allow nullable productId and handle null cases in deletion

+++ app/src/main/java/me/re/homestuffapp/domain/usecases/DeleteProduct.kt
@@ -6,7 +6,10 @@ import javax.inject.Inject
 class DeleteProduct @Inject constructor(
     private val productRepository: ProductRepositoryPort
 ) {
-    suspend operator fun invoke(productId: Long) {
+    suspend operator fun invoke(productId: Long?) {
+        if (productId == null) {
+            throw IllegalArgumentException("Product ID cannot be null")
+        }
         productRepository.deleteProduct(productId = productId)
     }
 }
